integrate limeuart into limepcie module, move drivers code into dedicated directory
