+++ drivers/CMakeLists.txt
@@ -1,0 +1,3 @@
+if(${CMAKE_SYSTEM_NAME} MATCHES "Linux")
+    add_subdirectory(linux)
+endif()
