+++ drivers/linux/limepcie/cmake/add_kernel_module.cmake
@@ -52,32 +52,19 @@ function(add_kernel_module)
         string(APPEND KBUILD_INCLUDE_DIR_FLAGS "-I${KBUILD_FILE_DIR}/${dir}")
     endforeach()
 
-    file(
-        WRITE ${KBUILD_FILE_DIR}/Kbuild
-        "ccflags-y += -Wno-declaration-after-statement
-    ccflags-y += -std=gnu11
-    ccflags-y += ${KBUILD_INCLUDE_DIR_FLAGS}
-    obj-m = ${KMOD_NAME}.o ")
-
     set(MODULE_KOBJECT ${KBUILD_FILE_DIR}/${KMOD_NAME}.ko)
     set(KERNEL_SOURCE_DIR /lib/modules/${KMOD_KERNEL_RELEASE}/build)
     set(KBUILD_CMD
         $(MAKE)
         -C
         ${KERNEL_SOURCE_DIR}
-        KMOD_ARCH=${KMOD_ARCH}
+        ARCH=${KMOD_ARCH}
         # Informs kbuild that an external module is being built.
         # The value given to "M" is the absolute path of the directory where the external module (kbuild file) is located.
         M=${KBUILD_FILE_DIR}
         modules)
 
-    set(KBUILD_CLEAN_CMD $(MAKE) -C ${KERNEL_SOURCE_DIR} KMOD_ARCH=${KMOD_ARCH} M=${KBUILD_FILE_DIR} clean)
-
-    set(KERNEL_SOURCE_RELATIVE_PATHS ${KMOD_SOURCES})
-    # Copy all source files into build directory and compile there, as the Kbuild produces artifacts in tree
-    foreach(SRC_FILENAME ${KERNEL_SOURCE_RELATIVE_PATHS})
-        configure_file(${CMAKE_CURRENT_LIST_DIR}/${SRC_FILENAME} ${KBUILD_FILE_DIR}/${SRC_FILENAME} COPYONLY)
-    endforeach()
+    set(KBUILD_CLEAN_CMD $(MAKE) -C ${KERNEL_SOURCE_DIR} ARCH=${KMOD_ARCH} M=${KBUILD_FILE_DIR} clean)
 
     add_custom_command(
         OUTPUT ${MODULE_KOBJECT}
@@ -96,6 +83,34 @@ function(add_kernel_module)
         install(FILES ${MODULE_KOBJECT} DESTINATION ${KERNEL_MODULE_DESTINATION})
         set(MODE_FLAG "--legacy")
     endif()
+
+    # Copy all source files into build directory and compile there, as the Kbuild produces artifacts in tree
+    set(OBJECT_FILES)
+    foreach(SRC_FILENAME ${KMOD_SOURCES})
+        configure_file(${CMAKE_CURRENT_LIST_DIR}/${SRC_FILENAME} ${KBUILD_FILE_DIR}/${SRC_FILENAME} COPYONLY)
+    endforeach()
+
+    # Collect output object files that should be linked into single module
+    # message(WARNING "SRC FILES: " "${KMOD_SOURCES}")
+    string(REGEX MATCHALL "[^ ;]+[.]c" KERNEL_OBJECTS_RELATIVE_PATHS "${KMOD_SOURCES}")
+    # message(WARNING "C FILES: ${KERNEL_OBJECTS_RELATIVE_PATHS}")
+    string(REGEX REPLACE "[.]c[;]?" ".o " KERNEL_OBJECTS_RELATIVE_PATHS "${KERNEL_OBJECTS_RELATIVE_PATHS}")
+    # message(WARNING "OBJ FILES: ${KERNEL_OBJECTS_RELATIVE_PATHS}")
+
+    # Kernel modules consisting of multiple source files cannot have name that matches o
+    list(FIND ${KERNEL_OBJECTS_RELATIVE_PATHS} "${KMOD_NAME}.c" SRC_FILENAME_MATCHES)
+    if (NOT ${SRC_FILENAME_MATCHES} EQUAL -1)
+        message(FATAL_ERROR "Kernel module consisting of multiple files cannot have name that matches source filename")
+    endif()
+
+    file(
+        WRITE ${KBUILD_FILE_DIR}/Kbuild
+        "ccflags-y += -Wno-declaration-after-statement
+    ccflags-y += -std=gnu11
+    ccflags-y += ${KBUILD_INCLUDE_DIR_FLAGS}
+    obj-m = ${KMOD_NAME}.o
+    ${KMOD_NAME}-y := ${KERNEL_OBJECTS_RELATIVE_PATHS}
+")
 
     # If module is already loaded, unload it
     install(CODE "execute_process(COMMAND ${CMAKE_CURRENT_LIST_DIR}/unload.sh ${KMOD_NAME})")
