+++ drivers/linux/limepcie/install.sh
