+++ drivers/linux/limepcie/limeuart.h
@@ -1,0 +1,7 @@
+#ifndef LIMEUART_H
+#define LIMEUART_H
+
+int limeuart_init(void);
+void limeuart_exit(void);
+
+#endif
