+++ drivers/linux/limepcie/uninstall-legacy-module.sh
