+++ drivers/linux/limepcie/uninstall.sh
